AIState, AIPathの管理をActorAIに移管 #2
BattleActorはAIをひとつまで持てる
AIを管理はするが、生成はしない

+++ game/ai_state_escape.cpp
@@ -49,7 +49,7 @@ void AIStateEscape::Update(void){
 		}
 	}
 
-	min_length = sqrt(min_length);
+	min_length = sqrtf(min_length);
 
 	const float SEARCH_RANGE(200.0f);
 	if(target_player == nullptr){return;}
